refactor(git): resolve code review findings - safety, structure, and quality
- Fix fail-open: recursion depth >4 now blocks instead of allowing
- Fix fail-open: catch block now exits 2 (deny) not 0 (allow)
- Add missing safety blocks: git gc --prune=now/all
- Extract git-utils.ts: consolidate 5x runGit, 2x getCurrentBranch, 2x isGitRepo
- Extract shell-tokenizer.ts: ~480 lines of tokenizer logic out of git-safety.ts
- Remove dead exports: stripQuotedContents and getStableRepoName with their tests
- Optimize session-summary.ts: replace read-concat-write with appendFile for JSONL
- Replace unsafe as-casts with in-operator narrowing in 4 type guard functions
- Fix fetch mock double-casts in event-bus-client.test.ts
- Merge examples.md into conventions.md (remove token duplication)
- Upgrade /git:history model from haiku to sonnet
- Add JSDoc to 4 exported functions missing documentation
- Add 20 todo files tracking resolved and deferred findings
