+++ plugins/git/commands/history.md
@@ -1,6 +1,6 @@
 ---
 description: Explore git commit history
-model: haiku
+model: sonnet
 allowed-tools: Bash(git log:*), Bash(git diff:*), Bash(git show:*), Bash(git status:*), Bash(git blame:*), Bash(git branch:*)
 argument-hint: [search-query]
 ---
