+++ todos/031-pending-p3-add-missing-jsdoc.md
@@ -1,0 +1,33 @@
+---
+status: pending
+priority: p3
+issue_id: "031"
+tags: [code-review, quality]
+dependencies: []
+---
+
+## Problem Statement
+
+Project convention says "NEVER write exported function without JSDoc". Several exported functions in the new git plugin lack JSDoc comments.
+
+## Findings
+
+The following exported functions have no JSDoc:
+
+- `git-safety.ts`: `checkCommand`, `checkFileEdit`, `isCommitCommand`, `getCurrentBranch`
+- `command-logger.ts`: `createLogEntry`
+- `session-summary.ts`: `extractFromTranscript`
+
+## Proposed Solutions
+
+Add JSDoc comments to all exported functions documenting what they do and why they exist.
+
+## Technical Details
+
+- **Files**: `plugins/git/hooks/git-safety.ts`, `plugins/git/hooks/command-logger.ts`, `plugins/git/hooks/session-summary.ts`
+- JSDoc should explain the "why" not just the "what", per project conventions
+
+## Acceptance Criteria
+
+- All exported functions in the git plugin have JSDoc comments
+- JSDoc includes a description of purpose and any non-obvious behavior
